crm_cnae/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
openpyxl
python-multipart
passlib[bcrypt]
python-jose
requests
beautifulsoup4